+++ tests/cli/qdeql/index.sh
@@ -8,7 +8,7 @@ root=$(realpath "${folder}/../../..")
 run_qdeql () {
     if [ "$#" -gt 1 ]
     then
-        printf '%s\0%s' "$(cat "$1")" "$2" | $TRILANGLE "${root}/qdeql/interpreter.trg"
+        printf '%s\0%s' "$(cat "$1")" "$2" | $TRILANGLE -af "${root}/qdeql/interpreter.trg"
     else
         $TRILANGLE "${root}/qdeql/interpreter.trg" <"$1"
     fi
@@ -27,3 +27,7 @@ text='
 
 output=$(run_qdeql "${folder}/cat.qd" "$text")
 test "$text" = "${output//$'\r'/}"
+
+# truth machine
+test 0 = "$(run_qdeql "${folder}/tm.qd" 0)"
+test 1111111111 = "$(run_qdeql "${folder}/tm.qd" 1 | head -c 10)"